backend/requirements.txt
django>=5.0.0
djangorestframework
django-cors-headers
celery
redis
psycopg2-binary
django-filter
django-extensions
